+++ src/device/tsp/Tsp/Unused/UnusedMethods.cs
@@ -33,7 +33,7 @@ internal static class UnusedMethods
         session.DeleteScript( scriptInfo.Title );
 
         SessionBaseExtensionMethods.TraceLastAction( $"\r\n\tImporting script from {scriptInfo.DeployFileFormat} '{deployFilePath}' file" );
-        session.ImportScript( scriptInfo.Title, deployFilePath, lineDelay );
+        session.ImportScript( scriptInfo.Compressor, scriptInfo.Encryptor, scriptInfo.Title, deployFilePath, lineDelay );
 
         session.RunScript( scriptInfo.Title, scriptInfo.VersionGetterElement );